refactor(application): convert to folder module, rename cog to ApplicationManagement
Moves modules/application.py, modules/conversations/application.py, and
modules/views/application.py into NerdyPy/modules/application/ as management.py,
conversations.py, and views.py. Removes the now-empty conversations/ and views/
directories. Updates all internal imports, bot.py view registration, utils/valkey.py,
and all test files to the new paths.
